refactor(openai): share Responses API helpers
